# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rollback Engine API",
    "version": "1.0.0",
    "description": "Event-sourced checkout, compensation, replay, recovery and debugging API. The Event Store is authoritative; snapshots and materialized views are derived optimizations."
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Local development server"
    }
  ],
...
  "paths": {
...
    "/orders": {
      "get": {
        "tags": ["Reads"],
        "operationId": "listOrders",
        "summary": "List orders from the materialized view",
        "description": "Optimized read. The materialized view can be unavailable or stale after a reported repair failure; replay endpoints remain authoritative.",
        "responses": {
          "200": {
            "description": "Active materialized orders",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Order" }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    },
    "/order": {
      "post": {
        "tags": ["Commands"],
        "operationId": "createOrder",
        "summary": "Create an order",
        "parameters": [
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CorrelationId" },
          { "$ref": "#/components/parameters/CausationId" }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateOrderRequest" }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Order created, or the original result for the same idempotency key",
            "headers": {
              "Idempotency-Key": { "$ref": "#/components/headers/IdempotencyKey" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Order" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "409": { "$ref": "#/components/responses/ConflictError" },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    },
    "/order/{id}": {
      "get": {
        "tags": ["Reads"],
        "operationId": "getOrder",
        "summary": "Get single order by aggregate ID",
        "parameters": [
          { "$ref": "#/components/parameters/OrderId" },
          {
            "name": "consistency",
            "in": "query",
            "required": false,
            "description": "Consistency mode for reading order state (materialized view vs authoritative replay)",
            "schema": {
              "type": "string",
              "enum": ["materialized", "authoritative"],
              "default": "materialized"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Order entity",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/OrderEntity" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      },
      "delete": {
        "tags": ["Commands"],
        "operationId": "deleteOrder",
        "summary": "Delete a compensated or inactive order",
        "parameters": [
          { "$ref": "#/components/parameters/OrderId" },
          { "$ref": "#/components/parameters/IdempotencyKey" },
          { "$ref": "#/components/parameters/CorrelationId" },
          { "$ref": "#/components/parameters/CausationId" }
        ],
        "responses": {
          "200": {
            "description": "Order deleted, or the original result for the same idempotency key",
            "headers": {
              "Idempotency-Key": { "$ref": "#/components/headers/IdempotencyKey" }
            },
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/DeleteOrderResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "409": { "$ref": "#/components/responses/ConflictError" },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    },
...
    "/history": {
      "get": {
        "tags": ["Reads", "Developer"],
        "operationId": "completeHistory",
        "summary": "Read the complete committed event history",
        "responses": {
          "200": {
            "description": "Global append order",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/HistoryResponse" }
              }
            }
          },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    },
    "/timeline/{orderId}": {
      "get": {
        "tags": ["Developer"],
        "operationId": "aggregateTimeline",
        "summary": "Read a compact aggregate trace",
        "description": "Derived only from committed events. Does not execute domain logic or create events.",
        "parameters": [
          { "$ref": "#/components/parameters/OrderIdByName" }
        ],
        "responses": {
          "200": {
            "description": "Sequence and trace metadata for one aggregate",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/EventTimeline" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    },
    "/replay-state/{orderId}": {
      "get": {
        "tags": ["Reads"],
        "operationId": "replayState",
        "summary": "Authoritatively rebuild aggregate state from events",
        "parameters": [
          { "$ref": "#/components/parameters/OrderIdByName" }
        ],
        "responses": {
          "200": {
            "description": "Full replay state",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectionState" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    },
    "/snapshot/{orderId}": {
      "get": {
        "tags": ["Reads"],
        "operationId": "readSnapshot",
        "summary": "Read the current snapshot optimization",
        "parameters": [
          { "$ref": "#/components/parameters/OrderIdByName" }
        ],
        "responses": {
          "200": {
            "description": "Stored snapshot",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["snapshot"],
                  "properties": {
                    "snapshot": { "$ref": "#/components/schemas/Snapshot" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    },
    "/replay-restore/{orderId}": {
      "post": {
        "tags": ["Recovery"],
        "operationId": "recoverMaterializedView",
        "summary": "Repair the materialized view from retained events",
        "description": "Recovery emits no domain events.",
        "parameters": [
          { "$ref": "#/components/parameters/OrderIdByName" }
        ],
        "responses": {
          "200": {
            "description": "Materialized view restored",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/RecoveryResponse" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    },
    "/state-at/{orderId}/{timestamp}": {
      "get": {
        "tags": ["Reads", "Developer"],
        "operationId": "stateAtTime",
        "summary": "Replay a contiguous event prefix at a timestamp",
        "parameters": [
          { "$ref": "#/components/parameters/OrderIdByName" },
          {
            "name": "timestamp",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "format": "date-time" }
          }
        ],
        "responses": {
          "200": {
            "description": "State at the requested time",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectionState" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    },
    "/state-at/{orderId}/sequence/{sequence}": {
      "get": {
        "tags": ["Reads", "Developer"],
        "operationId": "stateAtSequence",
        "summary": "Replay a contiguous event prefix up to an exact sequence number",
        "parameters": [
          { "$ref": "#/components/parameters/OrderIdByName" },
          {
            "name": "sequence",
            "in": "path",
            "required": true,
            "description": "Sequence number of the event prefix to project",
            "schema": { "type": "integer", "minimum": 0 }
          }
        ],
        "responses": {
          "200": {
            "description": "State at the requested sequence",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ProjectionState" }
              }
            }
          },
          "400": { "$ref": "#/components/responses/ValidationError" },
          "404": { "$ref": "#/components/responses/NotFoundError" },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    },
    "/diagnostics": {
      "get": {
        "tags": ["Developer"],
        "operationId": "getDiagnostics",
        "summary": "Query recorded engine diagnostics",
        "parameters": [
          {
            "name": "type",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "aggregateId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "commandId",
            "in": "query",
            "required": false,
            "schema": { "type": "string" }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1 }
          }
        ],
        "responses": {
          "200": {
            "description": "Diagnostic records",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "count": { "type": "integer" },
                    "diagnostics": {
                      "type": "array",
                      "items": { "type": "object" }
                    }
                  }
                }
              }
            }
          },
          "500": { "$ref": "#/components/responses/TechnicalError" }
        }
      }
    }
  },
  "components": {
    "parameters": {
      "IdempotencyKey": {
        "name": "Idempotency-Key",
        "in": "header",
        "required": false,
        "description": "Stable command identity. Reuse only with the same command type and normalized payload.",
        "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
      },
      "CorrelationId": {
        "name": "X-Correlation-Id",
        "in": "header",
        "required": false,
        "description": "Business-flow trace identifier shared across related commands.",
        "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
      },
      "CausationId": {
        "name": "X-Causation-Id",
        "in": "header",
        "required": false,
        "description": "Identifier that caused the first event. Following events point to the preceding event.",
        "schema": { "type": "string", "minLength": 1, "maxLength": 200 }
      },
      "OrderId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 }
      },
      "OrderIdByName": {
        "name": "orderId",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "minimum": 1 }
      }
    },
    "headers": {
      "IdempotencyKey": {
        "description": "Echoes a supplied Idempotency-Key after command context was accepted.",
        "schema": { "type": "string" }
      }
    },
    "responses": {
      "ValidationError": {
        "description": "Request validation failed. Correct the request before retrying.",
        "headers": {
          "Idempotency-Key": { "$ref": "#/components/headers/IdempotencyKey" }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "NotFoundError": {
        "description": "Aggregate, snapshot or route does not exist.",
        "headers": {
          "Idempotency-Key": { "$ref": "#/components/headers/IdempotencyKey" }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "ConflictError": {
        "description": "Domain, idempotency or optimistic-concurrency conflict. Inspect retryAction.",
        "headers": {
          "Idempotency-Key": { "$ref": "#/components/headers/IdempotencyKey" }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      },
      "TechnicalError": {
        "description": "Technical failure. Never infer retry safety from HTTP status alone.",
        "headers": {
          "Idempotency-Key": { "$ref": "#/components/headers/IdempotencyKey" }
        },
        "content": {
          "application/json": {
            "schema": { "$ref": "#/components/schemas/ErrorResponse" }
          }
        }
      }
    },
    "schemas": {
      "CreateOrderRequest": {
        "type": "object",
        "required": ["item"],
        "properties": {
          "item": { "type": "string", "minLength": 1 },
          "quantity": { "type": "integer", "minimum": 1, "default": 1 }
        }
      },
...
      "Order": {
        "type": "object",
        "required": ["id", "item", "quantity", "status", "createdAt"],
        "properties": {
          "id": { "type": "integer", "minimum": 1 },
          "item": { "type": "string" },
          "quantity": { "type": "integer", "minimum": 1 },
          "status": { "type": "string", "enum": ["created", "rolled_back"] },
          "createdAt": { "type": "string", "format": "date-time" },
          "rolledBackAt": { "type": "string", "format": "date-time" },
          "rollbackReason": { "type": ["string", "null"] }
        }
      },
      "EventMetadata": {
        "type": "object",
        "required": ["schemaVersion", "commandId", "correlationId", "causationId"],
        "properties": {
          "schemaVersion": { "type": "integer", "const": 1 },
          "commandId": { "type": ["string", "integer"] },
          "correlationId": { "type": ["string", "integer"] },
          "causationId": { "type": ["string", "integer"] }
        }
      },
      "DomainEvent": {
        "type": "object",
        "required": [
          "eventId",
          "eventType",
          "aggregateId",
          "sequence",
          "timestamp",
          "payload",
          "metadata"
        ],
        "properties": {
          "eventId": { "type": "string" },
          "eventType": {
            "type": "string",
            "enum": [
              "ORDER_CREATED",
              "INVENTORY_RESERVED",
              "PAYMENT_CHARGED",
              "PAYMENT_REFUNDED",
              "INVENTORY_RELEASED",
              "ORDER_ROLLED_BACK",
              "ORDER_DELETED"
            ]
          },
          "aggregateId": { "type": ["integer", "string"] },
          "sequence": { "type": "integer", "minimum": 1 },
          "timestamp": { "type": "string", "format": "date-time" },
          "payload": { "type": "object" },
          "metadata": { "$ref": "#/components/schemas/EventMetadata" }
        }
      },
      "ProjectionState": {
        "type": "object",
        "required": [
          "aggregateId",
          "version",
          "lifecycle",
          "deleted",
          "tombstone",
          "order",
          "inventory",
          "payment"
        ],
        "properties": {
          "aggregateId": { "type": ["integer", "string"] },
          "version": { "type": "integer", "minimum": 0 },
          "lifecycle": {
            "type": "string",
            "enum": ["empty", "active", "completed", "compensating", "rolled_back", "deleted"]
          },
          "deleted": { "type": "boolean" },
          "tombstone": { "type": ["object", "null"] },
          "order": {
            "oneOf": [
              { "$ref": "#/components/schemas/Order" },
              { "type": "null" }
            ]
          },
          "inventory": { "type": ["object", "null"] },
          "payment": { "type": ["object", "null"] }
        }
      },
      "Snapshot": {
        "type": "object",
        "required": ["aggregateId", "version", "timestamp", "state"],
        "properties": {
          "aggregateId": { "type": ["integer", "string"] },
          "version": { "type": "integer", "minimum": 1 },
          "timestamp": { "type": "string", "format": "date-time" },
          "state": { "$ref": "#/components/schemas/ProjectionState" }
        }
      },
      "SnapshotWarning": {
        "type": "object",
        "required": [
          "code",
          "category",
          "message",
          "eventCommitted",
          "retrySafe",
          "retryAction",
          "aggregateId"
        ],
        "properties": {
          "code": { "type": "string", "const": "SNAPSHOT_SAVE_FAILED" },
          "category": { "type": "string", "const": "technical" },
          "message": { "type": "string" },
          "eventCommitted": { "type": "boolean", "const": true },
          "retrySafe": { "type": "boolean", "const": false },
          "retryAction": { "type": "string", "const": "DO_NOT_RETRY_COMMAND" },
          "aggregateId": { "type": ["integer", "string"] }
        }
      },
...
      "DeleteOrderResponse": {
        "type": "object",
        "required": ["message", "order", "warnings"],
        "properties": {
          "message": { "type": "string" },
          "order": { "$ref": "#/components/schemas/Order" },
          "warnings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/SnapshotWarning" }
          }
        }
      },
      "HistoryResponse": {
        "type": "object",
        "required": ["count", "history"],
        "properties": {
          "count": { "type": "integer", "minimum": 0 },
          "history": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/DomainEvent" }
          }
        }
      },
      "TimelineEntry": {
        "type": "object",
        "required": [
          "sequence",
          "eventId",
          "eventType",
          "timestamp",
          "schemaVersion",
          "commandId",
          "correlationId",
          "causationId"
        ],
        "properties": {
          "sequence": { "type": "integer", "minimum": 1 },
          "eventId": { "type": "string" },
          "eventType": { "type": "string" },
          "timestamp": { "type": "string", "format": "date-time" },
          "schemaVersion": { "type": "integer" },
          "commandId": { "type": ["string", "integer"] },
          "correlationId": { "type": ["string", "integer"] },
          "causationId": { "type": ["string", "integer"] }
        }
      },
      "EventTimeline": {
        "type": "object",
        "required": [
          "aggregateId",
          "version",
          "eventCount",
          "commandIds",
          "correlationIds",
          "entries"
        ],
        "properties": {
          "aggregateId": { "type": ["integer", "string"] },
          "version": { "type": "integer", "minimum": 1 },
          "eventCount": { "type": "integer", "minimum": 1 },
          "commandIds": {
            "type": "array",
            "items": { "type": ["string", "integer"] }
          },
          "correlationIds": {
            "type": "array",
            "items": { "type": ["string", "integer"] }
          },
          "entries": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/TimelineEntry" }
          }
        }
      },
      "RecoveryResponse": {
        "type": "object",
        "required": ["orderId", "restored", "state"],
        "properties": {
          "orderId": { "type": ["integer", "string"] },
          "restored": { "type": "boolean", "const": true },
          "state": { "$ref": "#/components/schemas/ProjectionState" }
        }
      },
      "ErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "category",
          "message",
          "eventCommitted",
          "retrySafe",
          "retryAction"
        ],
        "properties": {
          "code": {
            "type": "string",
            "enum": [
              "VALIDATION_ERROR",
              "NOT_FOUND",
              "AGGREGATE_NOT_FOUND",
              "COMPENSATION_REQUIRED",
              "EVENT_COMMITTED_VIEW_REPAIR_FAILED",
              "IDEMPOTENCY_KEY_CONFLICT",
              "COMMAND_IN_PROGRESS",
              "COMMAND_EXECUTION_PARTIALLY_COMMITTED",
              "COMMAND_EXECUTION_INTERRUPTED_AFTER_COMMIT",
              "COMMAND_EVENT_HISTORY_INCONSISTENT",
              "COMMAND_STATE_PERSISTENCE_FAILED",
              "EVENT_APPEND_COMMIT_UNKNOWN",
              "COMMAND_RECONCILIATION_FAILED",
              "OPTIMISTIC_CONCURRENCY_CONFLICT",
              "INTERNAL_ERROR",
              "REQUEST_REJECTED"
            ]
          },
          "category": {
            "type": "string",
            "enum": ["validation", "not_found", "domain", "conflict", "technical", "request"]
          },
          "message": { "type": "string" },
          "eventCommitted": { "type": ["boolean", "null"] },
          "retrySafe": { "type": ["boolean", "null"] },
          "retryAction": {
            "type": "string",
            "enum": [
              "FIX_REQUEST",
              "COMPENSATE_THEN_RETRY",
              "USE_NEW_KEY",
              "WAIT_AND_RETRY_SAME_KEY",
              "RETRY_SAME_KEY",
              "RETRY_COMMAND",
              "REPLAY_SAME_KEY",
              "RECONCILE_SAME_KEY",
              "DO_NOT_RETRY_COMMAND",
              "MANUAL_RESOLUTION_REQUIRED"
            ]
          },
          "commandId": { "type": ["string", "integer"] },
          "aggregateId": { "type": ["string", "integer"] },
          "eventId": { "type": "string" },
          "eventIds": {
            "type": "array",
            "items": { "type": "string" }
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": { "$ref": "#/components/schemas/ErrorDetail" }
        }
      }
    }
  }
}